Final Develop

+++ HotelBooking.Web/Pages/RoomTypesBase.cs
@@ -1,6 +1,8 @@
 ﻿using HotelBooking.Shared.DTO;
 using HotelBooking.Web.Services.Interfaces;
 using Microsoft.AspNetCore.Components;
+using Microsoft.JSInterop;
+using static HotelBooking.Web.Pages.Message;
 
 namespace HotelBooking.Web.Pages
 {
@@ -10,26 +12,64 @@ namespace HotelBooking.Web.Pages
         public IEnumerable<RoomTypeDto> roomTypes;
         [Inject]
         public IRoomTypeService roomTypeService { get; set; }
-
+        [Inject] protected IJSRuntime JSRuntime { get; set; }
         public string ErrorMessage { get; set; }
         [Inject]
         public NavigationManager navigationManager { get; set; }
+        protected Message message = new Message();
+        public string messageText { get; set; }
+        public MessageType messageType { get; set; }
+
+        protected bool isElementHidden = true;
+
+        protected string GetElementStyle()
+        {
+            return isElementHidden ? "d-none" : string.Empty;
+        }
+        private void ShowSuccessMessage()
+        {
+            messageText = "Success Deleted";
+            messageType = MessageType.Success;
+
+            isElementHidden = false;
+        }
+        private void ShowErrorMessage(string message)
+        {
+            messageText = message;
+            messageType = MessageType.Error;
+            isElementHidden = false;
+        }
 
         protected override async Task OnInitializedAsync()
         {
-            roomTypes = await roomTypeService.GetRoomTypesAsync();
+            try
+            {
+
+                roomTypes = await roomTypeService.GetRoomTypesAsync();
+            }
+            catch(Exception ex)
+            {
+                ShowErrorMessage(ex.Message);
+            }
         }
 
         protected async Task DeleteCategory(int id)
         {
             try
             {
-                await roomTypeService.DeleteBookCategoryAsync(id);
-                navigationManager.NavigateTo("/RoomTypes", true);
+
+                bool confirmed = await JSRuntime.InvokeAsync<bool>("confirm", "Are you sure?");
+                if (confirmed)
+                {
+
+                    await roomTypeService.DeleteBookCategoryAsync(id);
+                    ShowSuccessMessage();
+                }
             }
             catch (Exception ex)
             {
                 ErrorMessage = ex.Message;
+                ShowErrorMessage(ex.Message);
             }
         }
     }
